+++ docs/source/releases.rst
@@ -26,6 +26,7 @@ Additions
 ~~~~~~~~~
 
 - Added :func:`qord.event` decorator for registering listeners in a subclassed :class:`Client`.
+- Added :meth:`Client.wait_for_event` method to allow waiting for event invocations.
 
 Bug fixes
 ~~~~~~~~~